refactor: expose aes to public.
