+++ src/joox/joox_v4_decryption.cpp
@@ -4,7 +4,7 @@
 #include "utils/paged_reader.h"
 #include "utils/pkcs7.hpp"
 
-#include <utils/aes.h>
+#include "parakeet-crypto/utils/aes.h"
 
 #include <cryptopp/modes.h>
 #include <cryptopp/pwdbased.h>
