+++ src/joox/joox_v4_encryption.cpp
@@ -3,7 +3,7 @@
 #include "utils/endian_helper.h"
 #include "utils/paged_reader.h"
 
-#include "utils/aes.h"
+#include "parakeet-crypto/utils/aes.h"
 
 #include <cryptopp/modes.h>
 #include <cryptopp/pwdbased.h>
