+++ src/ncm/ncm_key_utils.h
@@ -7,7 +7,7 @@
 #include <optional>
 #include <vector>
 
-#include <utils/aes.h>
+#include "parakeet-crypto/utils/aes.h"
 
 namespace parakeet_crypto::transformer
 {
